# contracts/kiln-http-api-v1.openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "Kiln HTTP API",
    "version": "1.0.0",
    "description": "Canonical inventory of every production HTTP route, method, transport, handler owner, request type, response type, status, media type, path/header/query input, and structured error envelope. Every named payload component delegates its complete field contract to a canonical generated schema."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8420"
    }
  ],
...
  "paths": {
...
    "/v1/corrections": {
      "delete": {
        "tags": [
          "corrections"
        ],
        "summary": "Clear corrections",
        "operationId": "corrections_clear_corrections_delete_v1_corrections",
        "x-kiln-handler": "corrections::clear_corrections",
        "x-kiln-rust-signature": "async fn clear_corrections( State(state): State<AppState>, ) -> Result<Json<ClearCorrectionsResponse>, ApiError>",
        "x-kiln-transport": "http",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearCorrectionsResponse"
                }
              }
            },
            "x-kiln-rust-type": "ClearCorrectionsResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "get": {
        "tags": [
          "corrections"
        ],
        "summary": "List corrections",
        "operationId": "corrections_list_corrections_get_v1_corrections",
        "x-kiln-handler": "corrections::list_corrections",
        "x-kiln-rust-signature": "async fn list_corrections( State(state): State<AppState>, Query(q): Query<ListQuery>, ) -> Json<ListResponse>",
        "x-kiln-transport": "http",
        "x-kiln-query-rust-type": "ListQuery",
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            },
            "x-kiln-rust-type": "ListResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      },
      "post": {
        "tags": [
          "corrections"
        ],
        "summary": "Upsert correction",
        "operationId": "corrections_upsert_correction_post_v1_corrections",
        "x-kiln-handler": "corrections::upsert_correction",
        "x-kiln-rust-signature": "async fn upsert_correction( State(state): State<AppState>, Json(row): Json<CorrectionRow>, ) -> Result<Json<CorrectionRow>, ApiError>",
        "x-kiln-transport": "http",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionRowInput"
              }
            }
          },
          "x-kiln-rust-type": "CorrectionRow"
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionRow"
                }
              }
            },
            "x-kiln-rust-type": "CorrectionRow"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/corrections/mark_trained": {
      "post": {
        "tags": [
          "corrections"
        ],
        "summary": "Mark trained",
        "operationId": "corrections_mark_trained_post_v1_corrections_mark_trained",
        "x-kiln-handler": "corrections::mark_trained",
        "x-kiln-rust-signature": "async fn mark_trained( State(state): State<AppState>, Json(req): Json<MarkTrainedRequest>, ) -> Result<Json<MarkTrainedResponse>, ApiError>",
        "x-kiln-transport": "http",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkTrainedRequest"
              }
            }
          },
          "x-kiln-rust-type": "MarkTrainedRequest"
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkTrainedResponse"
                }
              }
            },
            "x-kiln-rust-type": "MarkTrainedResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
    "/v1/corrections/{request_id}": {
      "delete": {
        "tags": [
          "corrections"
        ],
        "summary": "Delete correction",
        "operationId": "corrections_delete_correction_delete_v1_corrections_request_id",
        "x-kiln-handler": "corrections::delete_correction",
        "x-kiln-rust-signature": "async fn delete_correction( State(state): State<AppState>, AxumPath(request_id): AxumPath<String>, ) -> Result<Json<DeleteCorrectionResponse>, ApiError>",
        "x-kiln-transport": "http",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCorrectionResponse"
                }
              }
            },
            "x-kiln-rust-type": "DeleteCorrectionResponse"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "ApiError": {
        "description": "Structured API failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "hint"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "hint": {
                "type": "string"
              }
            }
          }
        },
        "x-kiln-rust-type": "ApiError",
        "x-kiln-field-schema-status": "complete"
      },
...
      "ClearCorrectionsResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/ClearCorrectionsResponse",
        "x-kiln-rust-type": "ClearCorrectionsResponse",
        "x-kiln-field-schema-status": "complete"
      },
...
      "CorrectionRow": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/CorrectionRow",
        "x-kiln-rust-type": "CorrectionRow",
        "x-kiln-field-schema-status": "complete"
      },
      "CorrectionRowInput": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/CorrectionRowInput",
        "x-kiln-rust-type": "CorrectionRow",
        "x-kiln-field-schema-status": "complete"
      },
...
      "DeleteCorrectionResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/DeleteCorrectionResponse",
        "x-kiln-rust-type": "DeleteCorrectionResponse",
        "x-kiln-field-schema-status": "complete"
      },
...
      "ListResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/ListResponse",
        "x-kiln-rust-type": "ListResponse",
        "x-kiln-field-schema-status": "complete"
      },
...
      "MarkTrainedRequest": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/MarkTrainedRequest",
        "x-kiln-rust-type": "MarkTrainedRequest",
        "x-kiln-field-schema-status": "complete"
      },
      "MarkTrainedResponse": {
        "$ref": "kiln-control-plane-v1.schema.json#/$defs/MarkTrainedResponse",
        "x-kiln-rust-type": "MarkTrainedResponse",
        "x-kiln-field-schema-status": "complete"
      },
...
    }
  },
...
}